+++ lib/ref_inspector.ex
@@ -22,15 +22,4 @@ defmodule RefInspector do
   """
   @spec reload() :: :ok
   defdelegate reload(), to: RefInspector.Database
-
-  @doc false
-  @spec reload_databases() :: :ok
-  def reload_databases() do
-    Logger.info(
-      "RefInspector.reload_databases/0 has been renamed to RefInspector.reload/0." <>
-        " The current alias will be removed in a future release."
-    )
-
-    reload()
-  end
 end